requirements.txt
streamlit>=1.28.0
PyPDF2>=3.0.1
sentence-transformers>=2.2.2
langchain>=0.0.350
faiss-cpu>=1.7.4
openai>=1.3.0
transformers>=4.35.0
torch>=2.1.0
numpy>=1.24.0
pickle5>=0.0.12
python-dotenv>=1.0.0
tiktoken>=0.5.1